+++ src/query.ts
@@ -1,38 +1,23 @@
 import { browser } from 'webextension-polyfill-ts';
 
-import {
-  DictType,
-  SearchRequest,
-  TranslateRequest,
-} from './background-request';
-import {
-  KanjiSearchResult,
-  NameResult,
-  NameSearchResult,
-  SearchResult,
-  WordSearchResult,
-} from './search-result';
-
-export type QueryResult = (
-  | WordSearchOrTranslateResult
-  | NameSearchResult
-  | KanjiSearchResult
-) & { preferNames: boolean };
-
-export interface WordSearchOrTranslateResult
-  extends Omit<WordSearchResult, 'matchLen'> {
-  type: 'words';
-  title?: string;
-  matchLen: number | null;
-}
+import { SearchRequest, TranslateRequest } from './background-request';
+import { NameResult, SearchResult, TranslateResult } from './search-result';
+import { stripFields } from './strip-fields';
 
 export interface QueryOptions {
   includeRomaji: boolean;
-  dict?: DictType;
-  prevDict: DictType | undefined;
-  preferNames: boolean;
   wordLookup: boolean;
 }
+
+export type QueryResult = SearchResult & {
+  title?: string;
+  namePreview?: NamePreview;
+};
+
+export type NamePreview = {
+  names: Array<NameResult>;
+  more: boolean;
+};
 
 type QueryCacheEntry = {
   hash: string;
@@ -49,9 +34,6 @@ export async function query(
   const hash = [
     text,
     options.includeRomaji ? '1' : '0',
-    options.dict || 'auto',
-    options.prevDict || 'none',
-    options.preferNames ? '1' : '0',
     options.wordLookup ? '1' : '0',
   ].join('-');
 
@@ -59,7 +41,7 @@ export async function query(
   // work out some sort of LRU scheme for removing entries. While there are
   // plenty of libraries for that and we even use one such in the background
   // script, this code is part of the content script which goes into every page
-  // so we try to keep it learn.
+  // so we try to keep it lean.
   //
   // As a result, we limit our cache size to 10 entries and just do a linear
   // search of the array.
@@ -82,10 +64,10 @@ export async function query(
   };
   const queryResult = doQuery(text, options)
     .then((result) => {
-      if (!result || (result.type === 'words' && result.dbUnavailable)) {
+      if (!result || !!result.dbStatus) {
         dropFromCache();
       }
-      return result;
+      return result ? addNamePreview(result) : null;
     })
     .catch(() => {
       dropFromCache();
@@ -106,25 +88,13 @@ async function doQuery(
   text: string,
   options: QueryOptions
 ): Promise<QueryResult | null> {
-  let message: SearchRequest | TranslateRequest;
-  if (options.wordLookup) {
-    message = {
-      type: 'search',
-      input: text,
-      dict: options.dict,
-      prevDict: options.prevDict,
-      preferNames: options.preferNames,
-      includeRomaji: options.includeRomaji,
-    };
-  } else {
-    message = {
-      type: 'translate',
-      title: text,
-      includeRomaji: options.includeRomaji,
-    };
-  }
+  const message: SearchRequest | TranslateRequest = {
+    type: options.wordLookup ? 'search' : 'translate',
+    input: text,
+    includeRomaji: options.includeRomaji,
+  };
 
-  let searchResult: SearchResult | null;
+  let searchResult: SearchResult | TranslateResult | null;
   try {
     searchResult = await browser.runtime.sendMessage(message);
   } catch (e) {
@@ -138,44 +108,97 @@ async function doQuery(
     return null;
   }
 
-  if (searchResult.type === 'kanji' || searchResult.type === 'names') {
-    return searchResult;
-  }
+  // Convert a translate result into a suitably shaped SearchResult but
+  // with the title part filled-in.
 
-  let matchLen: number | null = null;
-  if (searchResult.type === 'words') {
-    matchLen = searchResult.matchLen || 1;
-  }
-  const more = !!searchResult.more;
-
-  let title: string | undefined;
-  if (searchResult.type === 'translate') {
-    title = text.substr(0, searchResult.textLen);
+  let queryResult: QueryResult;
+  if (isTranslateResult(searchResult)) {
+    let title = text.substr(0, searchResult.textLen);
     if (text.length > searchResult.textLen) {
       title += '...';
     }
+    queryResult = {
+      words: {
+        ...stripFields(searchResult, ['dbStatus', 'textLen']),
+        type: 'words',
+        matchLen: searchResult.textLen,
+      },
+      title,
+      dbStatus: searchResult.dbStatus,
+    };
+  } else {
+    queryResult = searchResult;
   }
 
-  let names: Array<NameResult> | undefined;
-  let moreNames: boolean | undefined;
-  let preferNames = false;
-  let dbUnavailable = false;
-  if (searchResult.type === 'words') {
-    names = searchResult.names;
-    moreNames = searchResult.moreNames;
-    preferNames = searchResult.preferNames;
-    dbUnavailable = !!searchResult.dbUnavailable;
+  return queryResult;
+}
+
+function isTranslateResult(
+  result: SearchResult | TranslateResult
+): result is TranslateResult {
+  return (result as TranslateResult).type === 'translate';
+}
+
+function addNamePreview(result: QueryResult): QueryResult {
+  if (!result.words || !result.names) {
+    return result;
   }
+
+  // If we have a word result, check for a longer match in the names dictionary,
+  // but only if the existing match has some non-hiragana characters in it.
+  //
+  // The names dictionary contains mostly entries with at least some kanji or
+  // katakana but it also contains entries that are solely hiragana (e.g.  はなこ
+  // without any corresponding kanji). Generally we only want to show a name
+  // preview if it matches on some kanji or katakana as otherwise it's likely to
+  // be a false positive.
+  //
+  // While it might seem like it would be enough to check if the existing match
+  // from the words dictionary is hiragana-only, we can get cases where a longer
+  // match in the names dictionary _starts_ with hiragana but has kanji/katakana
+  // later, e.g. ほとけ沢.
+  const names: Array<NameResult> = [];
+  let more = false;
+
+  // Add up to three results provided that:
+  //
+  // - they have a kanji reading,
+  // - and are all are as long as the longest names match,
+  // - are all longer than the longest words match
+  const minLength = Math.max(result.names.matchLen, result.words.matchLen + 1);
+
+  for (const name of result.names.data) {
+    // Names should be in descending order of length so if any of them is less
+    // than the minimum length, we can skip the rest.
+    if (name.matchLen < minLength) {
+      break;
+    }
+
+    if (!name.k) {
+      continue;
+    }
+
+    if (names.length > 2) {
+      more = true;
+      break;
+    }
+
+    names.push(name);
+  }
+
+  if (!names.length) {
+    return result;
+  }
+
+  // If we got a match, extend the matchLen of the words result.
+  //
+  // Reaching into the words result like this is cheating a little bit but it
+  // simplifies the places where we use the word result.
+  const matchLen = names[0].matchLen;
 
   return {
-    type: 'words',
-    title,
-    names,
-    moreNames,
-    data: searchResult.data,
-    matchLen,
-    more,
-    preferNames,
-    dbUnavailable,
+    ...result,
+    words: { ...result.words, matchLen },
+    namePreview: { names, more },
   };
 }